optimize bitboard scoring function
fixed bug in the Position Tables (8x8 index but 10x12 tables), change
scoring data type from Double to Integer, multiply score with -1 only
once at the end to get black players score

+++ src/search/AdversarialSearch.java
@@ -19,5 +19,5 @@ public interface AdversarialSearch {
 	 * @param evalFunction the eval function that scores a leaf
 	 * @return Pair<T, Double> a pair (bestMove, score)
 	 */
-	Pair<Node, Double> search(Node start, Function<Node, Double> evalFunction);
+	Pair<Node, Integer> search(Node start, Function<Node, Integer> evalFunction);
 }
