+++ src/search/evalfunctions/ScoreBoard.java
@@ -4,6 +4,6 @@ import java.util.function.Function;
 
 import board.Board;
 
-public interface ScoreBoard extends Function<Board, Double> {
+public interface ScoreBoard extends Function<Board, Integer> {
 	public int getScoreCounter();
 }
