+++ src/search/functions/BoardFunction.java
@@ -5,15 +5,15 @@ import java.util.function.Function;
 import board.Board;
 import search.Node;
 
-public class BoardFunction implements Function<Node, Double> {
-	private Function<Board, Double> f;
+public class BoardFunction implements Function<Node, Integer> {
+	private Function<Board, Integer> f;
 
-	public BoardFunction(Function<Board, Double> f) {
+	public BoardFunction(Function<Board, Integer> f) {
 		this.f = f;
 	}
 
 	@Override
-	public Double apply(Node node) {
+	public Integer apply(Node node) {
 		Board board = (Board) node.getState();
 		return f.apply(board);
 	}
